Location: Remove all references to proprietary dependencies
These libraries are proprietary, and not guaranteed to always
be available at build time.
Actually, now they are never available at build time, since the
libraries are not included in the built system.img anymore.
Now, they are only found on the target device in a seperate partition.
Therefore, remove references that require build time linking.
For the headers, the extern specifiers are replaced with pointers in
preparation to load the symbols dynamically using libloc_loader.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>